% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_helpers.R
\name{add_dockerfile}
\alias{add_dockerfile}
\title{Create a Dockerfile for  Shiny App}
\usage{
add_dockerfile(input = "DESCRIPTION", output = "Dockerfile")
}
\arguments{
\item{input}{path to the DESCRIPTION file to use as an input.}

\item{output}{name of the Dockerfile output.}
}
\description{
Build a container containing your Shiny App.
}
\examples{
\dontrun{
add_dockerfile()
}
}
