% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dockerfiles_renv.R
\name{add_dockerfile_with_renv_}
\alias{add_dockerfile_with_renv_}
\title{add_dockerfile_with_renv_}
\usage{
add_dockerfile_with_renv_(
  path = ".",
  lockfile = NULL,
  output_dir = fs::path(tempdir(), "deploy"),
  distro = "focal",
  FROM = "rocker/verse",
  AS = NULL,
  sysreqs = TRUE,
  repos = c(CRAN = "https://cran.rstudio.com/"),
  expand = FALSE,
  extra_sysreqs = NULL,
  update_tar_gz = TRUE
)
}
\arguments{
\item{output_dir}{}
}
\description{
add_dockerfile_with_renv_
}
