% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_files.R
\name{add_rconnect_file}
\alias{add_rconnect_file}
\title{Add an app.R at the root of your package to deploy on RStudio Connect}
\usage{
add_rconnect_file(pkg = ".")
}
\arguments{
\item{pkg}{Path to the root of the package. Default is `"."`.}
}
\description{
Add an app.R at the root of your package to deploy on RStudio Connect
}
