% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_deploy_helpers.R
\name{add_shinyproxy_dockerfile}
\alias{add_shinyproxy_dockerfile}
\title{Create a Dockerfile for  Shiny App}
\usage{
add_shinyproxy_dockerfile(input = "DESCRIPTION", output = "Dockerfile",
  from = paste0("FROM rocker/tidyverse:", R.Version()$major, ".",
  R.Version()$minor))
}
\arguments{
\item{input}{path to the DESCRIPTION file to use as an input.}

\item{output}{name of the Dockerfile output.}

\item{from}{The FROM of the Dockerfile. Default is FROM rocker/tidyverse:
with `R.Version()$major` and `R.Version()$minor`.}
}
\description{
Build a container containing your Shiny App.
}
\examples{
\dontrun{
add_shinyproxy_dockerfile()
}
}
