% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem-yaml-utils.R
\name{amend_golem_config}
\alias{amend_golem_config}
\title{Amend golem config file}
\usage{
amend_golem_config(
  key,
  value,
  config = "default",
  golem_wd = golem::pkg_path(),
  talkative = TRUE,
  pkg
)
}
\arguments{
\item{key}{key of the value to add in \code{config}}

\item{value}{Name of value (\code{NULL} to read all values)}

\item{config}{Name of configuration to read from. Defaults to
the value of the \code{R_CONFIG_ACTIVE} environment variable
("default" if the variable does not exist).}

\item{golem_wd}{Path to the root of the package. Default is \code{get_golem_wd()}.}

\item{talkative}{Should the messages be printed to the console?}

\item{pkg}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This argument has been
replaced by \code{golem_wd} and is kept here for backward compatibility.
Providing a value to this argument has no effect: the value is silently
ignored, and \code{golem_wd} is used instead.}
}
\value{
Used for side effects.
}
\description{
Amend golem config file
}
