% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_attachment.R
\name{att_from_description}
\alias{att_from_description}
\title{Return all package dependencies from current package}
\usage{
att_from_description(path = "DESCRIPTION", dput = FALSE,
  field = c("Depends", "Imports", "Suggests"))
}
\arguments{
\item{path}{path to the DESCRIPTION file}

\item{dput}{if FALSE return a vector instead of dput output}

\item{field}{DESCRIPTION fied to parse, Import and Depends by default}
}
\value{
A character vector with packages names
}
\description{
Return all package dependencies from current package
}
