% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_button.R
\name{browser_button}
\alias{browser_button}
\title{Insert an hidden browser button}
\usage{
browser_button()
}
\value{
Prints the code to the console.
}
\description{
See \url{https://rtask.thinkr.fr/blog/a-little-trick-for-debugging-shiny/} for more context.
}
