% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing_ns.R
\name{check_namespace_sanity}
\alias{check_namespace_sanity}
\title{check namespace sanity
Will check if the namespace (NS) are correctly set in the shiny modules}
\usage{
check_namespace_sanity(
  pkg = golem::get_golem_wd(),
  extend_input_output_funmodule = NA_character_,
  auto_fix = TRUE
)
}
\arguments{
\item{pkg}{Character. The package path}

\item{extend_input_output_funmodule}{Character. Extend the input, output or function module to check}

\item{auto_fix}{Logical. Fix the missing namespace automatically. Default is TRUE}
}
\value{
Logical. TRUE if the namespace are correctly set, FALSE otherwise
}
\description{
check namespace sanity
Will check if the namespace (NS) are correctly set in the shiny modules
}
