% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_renv.R
\name{create_renv_for_dev}
\alias{create_renv_for_dev}
\alias{create_renv_for_prod}
\title{Title}
\usage{
create_renv_for_dev(
  path = ".",
  dev_pkg = c("renv", "devtools", "roxygen2", "usethis", "pkgload", "testthat", "covr",
    "attachment", "pak", "dockerfiler",
    "remotes::install_github('ThinkR-open/checkhelper')"),
  folder_to_include = c("dev/", "data-raw/"),
  output = "renv.lock"
)

create_renv_for_prod(path = ".", output = "renv.lock.prod")
}
\arguments{
\item{output}{}
}
\value{

}
\description{
Title
}
