% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_run_dev.R
\name{create_run_dev}
\alias{create_run_dev}
\title{Create run_dev by module}
\usage{
create_run_dev(name, module_ui, module_server, ..., pkg = get_golem_wd(),
  open = TRUE)
}
\arguments{
\item{name}{name of the script}

\item{module_ui}{name of the module UI}

\item{module_server}{name of the module server}

\item{...}{add parameters to your module_server. They have to be a string like param = "'jean'"}

\item{pkg}{path to package}

\item{open}{boolean open the file}
}
\description{
Allow you to create one run_dev by module
}
\examples{

\dontrun{
create_run_dev("ok", 
               mod_ok_ui,
               mod_ok_server, 
               r= "reactiveValues(message = 'ok')", 
               message = "'have to be string'")
}
}
