% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shiny_template.R
\name{create_shiny_template}
\alias{create_shiny_template}
\title{Create a package for Shiny App}
\usage{
create_shiny_template(path, check_name = TRUE, ...)
}
\arguments{
\item{path}{Name of the folder to create the package in. This will also be 
used as the package name.}

\item{check_name}{When using this function in the console, you can prevent 
the package name from being checked.}

\item{...}{not used}
}
\description{
Create a package for Shiny App
}
