% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disable_autoload.R
\name{disable_autoload}
\alias{disable_autoload}
\title{Disabling Shiny Autoload of R Scripts}
\usage{
disable_autoload(golem_wd = get_golem_wd(), pkg)
}
\arguments{
\item{golem_wd}{Path to the root of the package. Default is \code{get_golem_wd()}.}

\item{pkg}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This argument has been
replaced by \code{golem_wd} and is kept here for backward compatibility.
Providing a value to this argument has no effect: the value is silently
ignored, and \code{golem_wd} is used instead.}
}
\value{
The path to the file, invisibly.
}
\description{
Disabling Shiny Autoload of R Scripts
}
\examples{
if (interactive()) {
  disable_autoload()
}
}
