% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_deploy_helpers.R
\name{dock_from_desc}
\alias{dock_from_desc}
\title{Create Dockerfile from DESCRIPTION}
\usage{
dock_from_desc(path = "DESCRIPTION", FROM = paste0("rocker/r-ver:",
  R.Version()$major, ".", R.Version()$minor), AS = NULL,
  sysreqs = TRUE, repos = "https://cran.rstudio.com/",
  expand = FALSE, build_golem_from_source = FALSE,
  update_tar_gz = TRUE)
}
\arguments{
\item{path}{path to the DESCRIPTION file to use as an input.}

\item{FROM}{The FROM of the Dockerfile. Default is FROM rocker/r-ver:
with \code{R.Version()$major} and \code{R.Version()$minor}.}

\item{AS}{The AS of the Dockerfile. Default it NULL.}

\item{sysreqs}{boolean to check the system requirements}

\item{repos}{character vector, the base URL of the repositories}

\item{expand}{boolean, if \code{TRUE} each system requirement will be known his own RUN line}
}
\description{
Create Dockerfile from DESCRIPTION
}
