% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_r_files.R
\name{fct_template}
\alias{fct_template}
\title{Golem Function Template}
\usage{
fct_template(name, path, export = FALSE, ...)
}
\arguments{
\item{name}{The name of the generated function.}

\item{path}{The path to the R script where the function will be written.
Note that this path will not be set by the user but via
\code{add_fct()}.}

\item{export}{Whether the generated function should be exported.}

\item{...}{Extra arguments, ignored by the default template.}
}
\value{
Used for side effect
}
\description{
Function templates can be used to extend the \code{add_fct()} creation
mechanism with your own template, so that you can be even more
productive when building your \code{{shiny}} app.
Function template functions do not aim at being called as is by
users, but to be passed as an argument to the \code{add_fct()} function.
}
\details{
A template function can take the following arguments to be passed
from \code{add_fct()}:
\itemize{
\item name: the name of the function
\item path: the path to the file in R/
\item export: a TRUE/FALSE value
}

If you want your function to ignore these parameters, set \code{...} as
the last argument of your function, then these will be ignored. See
the examples section of this help.
}
\examples{
if (interactive()) {
  my_tmpl <- function(name, path, ...) {
    # Define a template that writes to the function file
    write(name, path)
  }
  golem::add_fct(name = "custom", template = my_tmpl)
}

}
\seealso{
\code{\link[=add_fct]{add_fct()}}
}
