% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_shinyproxy.R
\name{gen_dockerfile}
\alias{gen_dockerfile}
\title{Create a proper Dockerfile}
\usage{
gen_dockerfile(path = "DESCRIPTION", output = "Dockerfile")
}
\arguments{
\item{path}{path to DESCRIPTION file}

\item{output}{path to Dockerfile}
}
\description{
build a container containing your shiny app
}
\examples{
\dontrun{
gen_dockerfile()
}
}
