% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_current_config}
\alias{get_current_config}
\title{Return path to the \code{{golem}} config-file}
\usage{
get_current_config(path = getwd())
}
\arguments{
\item{path}{character string giving the path to start looking for the config;
the usual value is the \code{{golem}}-package top-level directory but a user
supplied config is now supported (see \strong{Details} for how to use this
feature).}
}
\value{
character string giving the path to the \code{{golem}} config-file
}
\description{
This function tries to find the current config file, being
either inst/golem-config.yml or the GOLEM_CONFIG_PATH env var
}
\details{
In most cases this function simply returns the path to the default
golem-config file located under "inst/golem-config.yml". That config comes
in \code{yml}-format, see the \href{https://engineering-shiny.org/golem.html?q=config#golem-config}{Engineering Production-Grade Shiny Apps}
for further details on its format and how to set options therein.

Advanced app developers may benefit from having an additional user
config-file. This is achieved with setting the GOLEM_CONFIG_PATH env var.
}
