% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dependencies.R
\name{get_dependencies}
\alias{get_dependencies}
\title{Return all package dependencies from current package}
\usage{
get_dependencies(path = "DESCRIPTION", pkg = get_golem_wd(),
  dput = FALSE, field = c("Imports"))
}
\arguments{
\item{path}{path to the DESCRIPTION file}

\item{pkg}{Path to the root of the package. Default is \code{"."}.}

\item{dput}{if TRUE return a dput output instead of character vector}

\item{field}{DESCRIPTION fields to parse. Default is Import}
}
\description{
Return all package dependencies from current package
}
\examples{
\donttest{
if (interactive()){
  get_dependencies()
}
}
}
