% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dev.R
\name{get_files}
\alias{get_files}
\title{Get hash to now if something change}
\usage{
get_files(dirs)
}
\arguments{
\item{dirs}{vector of dirs to check in package}
}
\value{
character string
}
\description{
Get hash to now if something change
}
