% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{go_to}
\alias{go_to}
\title{{golem} addins}
\usage{
go_to(file, wd = golem::get_golem_wd())
}
\arguments{
\item{file}{Name of the file to be loaded from the \code{dev} directory.}

\item{wd}{Working directory of the current golem package.}
}
\description{
Addins to go to common files used in developing a \code{{golem}} package.
}
