% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem-package.R
\docType{package}
\name{golem}
\alias{golem-package}
\alias{golem}
\title{A package for building Shiny App}
\description{
Read more about building big shiny apps at \url{https://engineering-shiny.org/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://thinkr-open.github.io/golem/}
  \item \url{https://github.com/ThinkR-open/golem}
  \item Report bugs at \url{https://github.com/ThinkR-open/golem/issues}
}

}
\author{
\strong{Maintainer}: Colin Fay \email{contact@colinfay.me} (\href{https://orcid.org/0000-0001-7343-1846}{ORCID})

Authors:
\itemize{
  \item Colin Fay \email{contact@colinfay.me} (\href{https://orcid.org/0000-0001-7343-1846}{ORCID})
  \item Vincent Guyader \email{vincent@thinkr.fr} (\href{https://orcid.org/0000-0003-0671-9270}{ORCID}) (previous maintainer)
  \item Sébastien Rochette \email{sebastien@thinkr.fr} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})
  \item Cervan Girard \email{cervan@thinkr.fr} (\href{https://orcid.org/0000-0002-4816-4624}{ORCID})
}

Other contributors:
\itemize{
  \item Novica Nakov \email{nnovica@gmail.com} [contributor]
  \item David Granjon \email{dgranjon@ymail.com} [contributor]
  \item Arthur Bréant \email{arthur@thinkr.fr} [contributor]
  \item Antoine Languillaume \email{antoine@thinkr.fr} [contributor]
  \item Ilya Zarubin \email{zarubin@wiso.uni-koeln.de} [contributor]
  \item ThinkR [copyright holder]
}

}
