% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_golem_name}
\alias{get_golem_name}
\alias{app_sys}
\title{A function to return the golem name to be used elsewhere
in the package.}
\usage{
get_golem_name()

app_sys()
}
\value{
The name of the golem.

The path to the current golem.
}
\description{
This function allows to use \code{app_sys()} instead of using the
\code{system.file(, package = "pkg")}.
}
