% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem-yaml-get.R, R/golem-yaml-set.R,
%   R/set_golem_options.R
\name{get_golem_wd}
\alias{get_golem_wd}
\alias{get_golem_name}
\alias{get_golem_version}
\alias{set_golem_name}
\alias{set_golem_version}
\alias{set_golem_options}
\title{\code{{golem}} options}
\usage{
get_golem_wd(use_parent = TRUE, golem_wd = golem::pkg_path(), pkg)

get_golem_name(
  config = Sys.getenv("GOLEM_CONFIG_ACTIVE", Sys.getenv("R_CONFIG_ACTIVE", "default")),
  use_parent = TRUE,
  golem_wd = golem::pkg_path(),
  pkg
)

get_golem_version(
  config = Sys.getenv("GOLEM_CONFIG_ACTIVE", Sys.getenv("R_CONFIG_ACTIVE", "default")),
  use_parent = TRUE,
  golem_wd = golem::pkg_path(),
  pkg
)

set_golem_name(
  name = golem::pkg_name(),
  golem_wd = golem::pkg_path(),
  talkative = TRUE,
  old_name = golem::pkg_name(),
  pkg
)

set_golem_version(
  version = golem::pkg_version(),
  golem_wd = golem::pkg_path(),
  talkative = TRUE,
  pkg
)

set_golem_options(
  golem_name = golem::pkg_name(),
  golem_version = golem::pkg_version(),
  golem_wd = golem::pkg_path(),
  app_prod = FALSE,
  talkative = TRUE,
  config_file = golem::get_current_config(golem_wd)
)
}
\arguments{
\item{use_parent}{\code{TRUE} to scan parent directories for
configuration files if the specified config file isn't found.}

\item{golem_wd}{Working directory of the current golem package.}

\item{pkg}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This argument has been
replaced by \code{golem_wd} and is kept here for backward compatibility.
Providing a value to this argument has no effect: the value is silently
ignored, and \code{golem_wd} is used instead.}

\item{config}{Name of configuration to read from. Defaults to
the value of the \code{R_CONFIG_ACTIVE} environment variable
("default" if the variable does not exist).}

\item{name}{The name of the app}

\item{talkative}{Should the messages be printed to the console?}

\item{old_name}{The old name of the app, used when changing the name}

\item{version}{The version of the app}

\item{golem_name}{Name of the current golem.}

\item{golem_version}{Version of the current golem.}

\item{app_prod}{Is the \code{{golem}} in prod mode?}

\item{config_file}{path to the \code{{golem}} config file}
}
\value{
Used for side-effects for the setters, and values from the
config in the getters.
}
\description{
Set and get a series of options to be used with \code{{golem}}.
These options are found inside the \code{golem-config.yml} file, found in most cases
inside the \code{inst} folder.
}
\section{Set Functions}{

\itemize{
\item \code{set_golem_options()} sets all the options, with the defaults from the functions below.
\item \code{set_golem_wd()} defaults to \code{golem::golem_wd()}, which is the package root when starting a golem.
\item \code{set_golem_name()} defaults \code{golem::pkg_name()}
\item \code{set_golem_version()} defaults \code{golem::pkg_version()}
}
}

\section{Get Functions}{

Reads the information from \code{golem-config.yml}
\itemize{
\item \code{get_golem_wd()}
\item \code{get_golem_name()}
\item \code{get_golem_version()}
}
}

