% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_golem_wd}
\alias{get_golem_wd}
\alias{set_golem_wd}
\title{Get and set \code{{golem}} working directory}
\usage{
get_golem_wd()

set_golem_wd(path, talkative = TRUE)
}
\arguments{
\item{path}{The path to set the golem working directory.
Note that it will be passed to \code{normalizePath}.}

\item{talkative}{Should the function print where the
new path is defined?}
}
\value{
The path to the working directory.
}
\description{
Many \code{{golem}} functions rely on a specific working directory,
most of the time the root of the package. This working directory
is set by \code{set_golem_options} or the first time you create a file.
It default to \code{"."}, the current directory when starting a golem.
You can use these two functions if you need to manipulate this
directory.
}
