% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dev.R
\name{golemLaunch}
\alias{golemLaunch}
\title{golemLaunch}
\value{
methods to manipulate your run_dev
}
\description{
golemLaunch

golemLaunch
}
\details{
A tools to use run_dev scripts
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{process}}{name of the process}

\item{\code{dir_of_app}}{app dir}

\item{\code{run_dev}}{loop for restart}

\item{\code{hash}}{hash to now save time of files}

\item{\code{path_to_run_dev}}{to the run_dev script}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{golemLaunch$new()}}
\item \href{#method-start}{\code{golemLaunch$start()}}
\item \href{#method-stop}{\code{golemLaunch$stop()}}
\item \href{#method-restart}{\code{golemLaunch$restart()}}
\item \href{#method-is_running}{\code{golemLaunch$is_running()}}
\item \href{#method-auto_reload}{\code{golemLaunch$auto_reload()}}
\item \href{#method-stop_autoreload}{\code{golemLaunch$stop_autoreload()}}
\item \href{#method-clone}{\code{golemLaunch$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new object to interact with the dev app
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{golemLaunch$new(file = "dev/run_dev.R", path = golem::get_golem_wd())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{path of the script to run}

\item{\code{path}}{path to the app folder initialize with golem, by default get_golem_wd}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object to interact with dev app
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start"></a>}}
\if{latex}{\out{\hypertarget{method-start}{}}}
\subsection{Method \code{start()}}{
Start the app
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{golemLaunch$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\if{latex}{\out{\hypertarget{method-stop}{}}}
\subsection{Method \code{stop()}}{
Stop the app
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{golemLaunch$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-restart"></a>}}
\if{latex}{\out{\hypertarget{method-restart}{}}}
\subsection{Method \code{restart()}}{
Restart the app
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{golemLaunch$restart()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_running"></a>}}
\if{latex}{\out{\hypertarget{method-is_running}{}}}
\subsection{Method \code{is_running()}}{
Find if the process of the app is running
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{golemLaunch$is_running()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-auto_reload"></a>}}
\if{latex}{\out{\hypertarget{method-auto_reload}{}}}
\subsection{Method \code{auto_reload()}}{
Auto reload shiny app develop with {golem}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{golemLaunch$auto_reload()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop_autoreload"></a>}}
\if{latex}{\out{\hypertarget{method-stop_autoreload}{}}}
\subsection{Method \code{stop_autoreload()}}{
Stop auto reload
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{golemLaunch$stop_autoreload()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{golemLaunch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
