% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dev_deps.R
\name{install_dev_deps}
\alias{install_dev_deps}
\title{Install \code{{golem}} dev dependencies}
\usage{
install_dev_deps(dev_deps, force_install = FALSE, ...)
}
\arguments{
\item{dev_deps}{optional character vector of packages to install}

\item{force_install}{If force_install is TRUE,
then the user is not interactively asked
to install them.}

\item{...}{further arguments passed to the install function.}
}
\value{
Used for side-effects
}
\description{
This function will run rlang::check_installed() on:
\itemize{
\item \code{usethis}
\item \code{pkgload}
\item \code{dockerfiler}
\item \code{devtools}
\item \code{roxygen2}
\item \code{attachment}
\item \code{rstudioapi}
\item \code{fs}
\item \code{desc}
\item \code{pkgbuild}
\item \code{processx}
\item \code{rsconnect}
\item \code{testthat}
}
}
\examples{
if (interactive()) {
  install_dev_deps()
}

}
