% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_golem.R
\name{is_golem}
\alias{is_golem}
\title{Is the directory a golem-based app?}
\usage{
is_golem(path = getwd())
}
\arguments{
\item{path}{Path to the directory to check.
Defaults to the current working directory.}
}
\value{
A boolean, \code{TRUE} if the directory is a golem-based app, \code{FALSE} else.
}
\description{
Trying to guess if \code{path} is a golem-based app.
}
\examples{
is_golem()
}
