% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_running.R
\name{is_running}
\alias{is_running}
\title{Is the running app a golem app?}
\usage{
is_running()
}
\value{
TRUE if the running app is a \code{{golem}} based app,
FALSE otherwise.

A boolean.
}
\description{
Note that this will return \code{TRUE} only if the application
has been launched with \code{with_golem_options()}
}
\examples{
is_running()
}
