% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js.R
\name{js}
\alias{js}
\title{Add JS code Straight to your app (instead of a dependency)}
\usage{
js()
}
\value{
A script
}
\description{
This function can be used in your UI to insert directly the JavaScript 
functions contained in golem.
}
\details{
These functions are meant to be used with `session$sendCustomMessage` 
    from the server side. 

\describe{
  \item{show}{Show an element with the jQuery selector provided.}
  \item{hide}{Hide an element with the jQuery selector provided.}
  \item{showid}{Show an element with the id provided.}
  \item{hideid}{Hide an element with the id provided.}
  \item{showclass}{Same as showid, but with class.}
  \item{hideclass}{Same as hideid, but with class.}
  \item{showhref}{Same as showid, but with `a[href*=`}
  \item{hidehref}{Same as hideid, but with `a[href*=`}
  \item{clickon}{Click on an element. The full jQuery selector has to be used.}
  \item{disable}{Add "disabled" to an element. The full jQuery selector has to be used.}
  \item{reable}{Remove "disabled" from an element. The full jQuery selector has to be used.}
}
}
