% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dev.R
\docType{data}
\name{loops}
\alias{loops}
\title{loops env for autoreload}
\format{
An object of class \code{environment} of length 0.
}
\usage{
loops
}
\description{
loops env for autoreload
}
\keyword{datasets}
