% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintenance_page.R
\name{maintenance_page}
\alias{maintenance_page}
\title{maintenance_page}
\usage{
maintenance_page()
}
\value{
an html_document
}
\description{
A default html page for maintenance mode
}
\details{
see the vignette \code{vignette("f_extending_golem", package = "golem")} for details.
}
