% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dev.R
\name{make_dev}
\alias{make_dev}
\title{Make a function dependent to dev mode}
\usage{
make_dev(fun)
}
\arguments{
\item{fun}{A function}
}
\description{
The function returned will be run only if `golem::app_dev()`
    returns TRUE.
}
