% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_modules.R
\name{moduleServer}
\alias{moduleServer}
\title{shim for moduleServer}
\usage{
moduleServer(id, module, ..., session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{module}{A Shiny module server function.}

\item{...}{Arguments to pass to the module server function when it's defined outside of \code{moduleServer()}}

\item{session}{Session from which to make a child scope (the default should
almost always be used).}
}
\value{
The return value, if any, from executing the module server function
}
\description{
This function is a shim for \code{shiny::moduleServer} that allows passing
arguments via \code{...}
}
\examples{
name_ui <- function(id){
ns <- NS(id)
tagList(
  actionButton(ns("go"), "go")
)
}

name_server_core <- function(input, output, session, arg) {
  observeEvent( input$go , {
    print(arg)
  })
}

name_server <- function(id, arg) {
  moduleServer(
   id,
   name_server_core, 
   arg = arg
  )
}
}
