% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dev.R
\name{app_prod}
\alias{app_prod}
\alias{app_dev}
\title{Is the app in dev mode or prod mode?}
\usage{
app_prod()

app_dev()
}
\value{
\code{TRUE} or \code{FALSE} depending on the status of \code{getOption( "golem.app.prod")}

A boolean.
}
\description{
Is the app in dev mode or prod mode?
}
