% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_golem_options}
\alias{set_golem_options}
\title{{golem} options}
\usage{
set_golem_options(pkg_name = pkgload::pkg_name(),
  pkg_version = pkgload::pkg_version(), pkg_path = pkgload::pkg_path(),
  app_prod = FALSE)
}
\arguments{
\item{pkg_name}{Name of the current golem.}

\item{pkg_version}{Version of the current golem.}

\item{pkg_path}{Name of the current golem.}

\item{app_prod}{Is the {golem} in prod mode?}
}
\description{
Set a series of options to be used internally by {golem}.
}
