% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem-yaml-set.R
\name{set_golem_wd}
\alias{set_golem_wd}
\title{Set the golem working directory}
\usage{
set_golem_wd(
  new_golem_wd = golem::pkg_path(),
  current_golem_wd = golem::pkg_path(),
  talkative = TRUE,
  golem_wd,
  pkg
)
}
\arguments{
\item{new_golem_wd, current_golem_wd}{New & current directory,
to be used in \code{set_golem_wd()}.}

\item{talkative}{Should the messages be printed to the console?}

\item{golem_wd}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This argument has been
replaced by \code{new_golem_wd} and is kept here for backward compatibility.
Providing a value to this argument has no effect: the value is silently
ignored, and \code{new_golem_wd} is used instead.}

\item{pkg}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This argument has been
replaced by \code{current_golem_wd} and is kept here for backward
compatibility. Providing a value to this argument has no effect: the
value is silently ignored, and \code{current_golem_wd} is used instead.}
}
\value{
Used for side-effects, and returns the \code{golem_wd} path invisibly.
}
\description{
Set the \code{golem_wd} value (the working directory of the current golem
package) inside the \code{golem-config.yml} file.
}
