% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{js_handler_template}
\alias{js_handler_template}
\alias{js_template}
\alias{css_template}
\alias{sass_template}
\alias{blank_template}
\title{Golem's default custom templates}
\usage{
js_handler_template(path, name = "fun", code = " ")

js_template(path, code = " ")

css_template(path, code = " ")

sass_template(path, code = " ")

blank_template(path, code = " ")
}
\arguments{
\item{path}{The path to the JS script where this template will be written.}

\item{name}{Shiny's custom handler name.}

\item{code}{JavaScript code to be written in the function.}
}
\value{
Used for side effect
}
\description{
These functions do not aim at being called as is by
users, but to be passed as an argument to the \code{add_js_handler()}
function.
}
\seealso{
\code{\link[=add_js_handler]{add_js_handler()}}
}
