% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_helpers.R
\name{add_dockerfile}
\alias{add_dockerfile}
\title{Create a proper Dockerfile}
\usage{
add_dockerfile(path = "DESCRIPTION", output = "Dockerfile")
}
\arguments{
\item{from}{path to the DESCRIPTION file.}

\item{filename}{path to Dockerfile}
}
\description{
build a container containing your shiny app
}
\examples{
\dontrun{
add_dockerfile()
}
}
