% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_files.R
\name{add_js_file}
\alias{add_js_file}
\alias{add_js_handler}
\alias{add_css_file}
\title{Create a module}
\usage{
add_js_file(name, pkg = ".", dir = "inst/app/www", open = TRUE)

add_js_handler(name, pkg = ".", dir = "inst/app/www", open = TRUE)

add_css_file(name, pkg = ".", dir = "inst/app/www", open = TRUE)
}
\arguments{
\item{name}{The name of the module}

\item{pkg}{Path to the root of the package. Default is `"."`.}

\item{dir}{Path to the dir of the file}

\item{open}{Open the file}
}
\description{
These functions create files inside the `inst/app/www` folder.
}
