% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_files.R
\name{add_js_file}
\alias{add_js_file}
\alias{add_js_handler}
\alias{add_js_binding}
\alias{add_css_file}
\alias{add_ui_server_files}
\title{Create Files}
\usage{
add_js_file(
  name,
  pkg = get_golem_wd(),
  dir = "inst/app/www",
  open = TRUE,
  dir_create = TRUE,
  with_doc_ready = TRUE
)

add_js_handler(
  name,
  pkg = get_golem_wd(),
  dir = "inst/app/www",
  open = TRUE,
  dir_create = TRUE
)

add_js_binding(
  name,
  pkg = get_golem_wd(),
  dir = "inst/app/www",
  open = TRUE,
  dir_create = TRUE,
  initialize = FALSE,
  rate_policy = FALSE
)

add_css_file(
  name,
  pkg = get_golem_wd(),
  dir = "inst/app/www",
  open = TRUE,
  dir_create = TRUE
)

add_ui_server_files(pkg = get_golem_wd(), dir = "inst/app", dir_create = TRUE)
}
\arguments{
\item{name}{The name of the module}

\item{pkg}{Path to the root of the package. Default is \code{get_golem_wd()}.}

\item{dir}{Path to the dir where the file while be created.}

\item{open}{Should the file be opened?}

\item{dir_create}{Creates the directory if it doesn't exist, default is \code{TRUE}.}

\item{with_doc_ready}{Should the default file include \verb{$( document ).ready()}?}

\item{initialize}{Whether to add the initialize method. Default to FALSE.}

\item{rate_policy}{Whether to apply rate policy to callback method. Default
to FALSE. If TRUE, the debounce method with a delay of 250 ms is applied. YOu may
edit manually according to \url{https://shiny.rstudio.com/articles/building-inputs.html}.}
}
\description{
These functions create files inside the \code{inst/app} folder.
These functions can be used outside of a {golem} project.
}
