% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_files.R
\name{add_js_file}
\alias{add_js_file}
\alias{add_js_handler}
\alias{add_css_file}
\title{Create a module}
\usage{
add_js_file(name, pkg = ".")

add_js_handler(name, pkg = ".")

add_css_file(name, pkg = ".")
}
\arguments{
\item{name}{The name of the module}

\item{pkg}{Path to the root of the package. Default is `"."`.}
}
\description{
These functions create files inside the `inst/app/www` folder.
}
