% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ci_files.R
\name{add_github_action}
\alias{add_github_action}
\title{Add deployment CI for GitHub Actions}
\usage{
add_github_action(golem_wd = get_golem_wd(), open = TRUE)
}
\arguments{
\item{golem_wd}{Path to the root of the package. Default is \code{get_golem_wd()}.}

\item{open}{Should the created file be opened?}
}
\value{
The path to the created workflow, invisibly.
}
\description{
Creates a minimal GitHub Actions workflow for deploying a \code{{golem}} app to
Posit Connect via \code{{rsconnect}}. If needed, this function also creates a
root \code{app.R} and \code{.rscignore} by calling \code{\link[=add_positconnect_file]{add_positconnect_file()}}. The
generated Posit Connect entrypoint uses \code{{pkgload}}, so \code{{pkgload}} is added
to \code{DESCRIPTION}.
}
