% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ci_files.R
\name{add_gitlab_ci}
\alias{add_gitlab_ci}
\title{Add deployment CI for GitLab}
\usage{
add_gitlab_ci(golem_wd = get_golem_wd(), open = TRUE)
}
\arguments{
\item{golem_wd}{Path to the root of the package. Default is \code{get_golem_wd()}.}

\item{open}{Should the created file be opened?}
}
\value{
The path to the created GitLab CI file, invisibly.
}
\description{
Creates a minimal GitLab CI file for deploying a \code{{golem}} app to Posit
Connect via \code{{rsconnect}}. If needed, this function also creates a root
\code{app.R} and \code{.rscignore} by calling \code{\link[=add_positconnect_file]{add_positconnect_file()}}. The
generated Posit Connect entrypoint uses \code{{pkgload}}, so \code{{pkgload}} is added
to \code{DESCRIPTION}.
}
