% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_modules.R
\name{add_module}
\alias{add_module}
\title{Create a module}
\usage{
add_module(name, pkg = get_golem_wd(), open = TRUE,
<<<<<<< HEAD
  dir_create = TRUE, fct = NULL, utils = NULL)
=======
  dir_create = TRUE)
>>>>>>> start with issue #171
}
\arguments{
\item{name}{The name of the module}

\item{pkg}{Path to the root of the package. Default is \code{"."}.}

\item{open}{Should the file be opened?}

\item{dir_create}{Creates the directory if it doesn't exist, default is \code{TRUE}.}
<<<<<<< HEAD

\item{fct}{The name of the fct file.}

\item{utils}{The name of the utils file.}
=======
>>>>>>> start with issue #171
}
\description{
This function creates a module inside the \code{R/} folder, based
on a specific module structure.
}
\note{
This function will prefix the \code{name} argument with \code{mod_}.
}
