% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_golem.R
\name{create_golem}
\alias{create_golem}
\title{Create a package for Shiny App using golem}
\usage{
create_golem(
  path,
  check_name = TRUE,
  open = TRUE,
  package_name = basename(path),
  without_comments = FALSE,
  project_hook = golem::project_hook,
  ...
)
}
\arguments{
\item{path}{Name of the folder to create the package in. This will also be
used as the package name.}

\item{check_name}{When using this function in the console, you can prevent
the package name from being checked.}

\item{open}{Boolean open the created project}

\item{package_name}{Package name to use.By default it's \code{basename(path)} but if path == '.' and \code{package_name}
not explicitly given, then \code{basename(getwd())} will be used.}

\item{without_comments}{Boolean start project without golem comments}

\item{project_hook}{A function executed as a hook after project creation. Can be used to change the default \code{{golem}} structure.
to override the files and content. This function is executed}

\item{...}{Arguments passed to the \code{project_hook()} function.}
}
\description{
Create a package for Shiny App using golem
}
