% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_golem.R
\name{create_golem}
\alias{create_golem}
\title{Create a package for Shiny App using golem}
\usage{
create_golem(path, check_name = TRUE, open = TRUE,
  package_name = basename(path), without_comments = FALSE, ...)
}
\arguments{
\item{path}{Name of the folder to create the package in. This will also be
used as the package name.}

\item{check_name}{When using this function in the console, you can prevent
the package name from being checked.}

<<<<<<< HEAD
\item{open}{boolean open the created project}
=======
\item{open}{booleen open the created project}
>>>>>>> start with issue #171

\item{package_name}{package name to use}

\item{without_comments}{boolean start project without golem comments}

\item{...}{not used}
}
\description{
Create a package for Shiny App using golem
}
