% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shiny_template.R
\name{create_shiny_template}
\alias{create_shiny_template}
\title{Create a Rstudio Project dedicated to shinyapp}
\usage{
create_shiny_template(path, ...)
}
\arguments{
\item{path}{path to create}

\item{...}{not used}
}
\description{
Create a Rstudio Project dedicated to shinyapp
}
