% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disable_autoload.R
\name{disable_autoload}
\alias{disable_autoload}
\title{Disabling Shiny Autoload of R Scripts}
\usage{
disable_autoload(pkg = get_golem_wd())
}
\arguments{
\item{pkg}{Path to the root of the package. Default is \code{get_golem_wd()}.}
}
\value{
The path to the file, invisibly.
}
\description{
This function creates a module inside the \verb{R/} folder, based
on a specific module structure. This function can be used outside
of a \code{{golem}} project.
}
\examples{
if (interactive()){
  disable_autoload()
}
}
\seealso{
\code{\link[=module_template]{module_template()}}
}
