% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disable_autoload.R
\name{disable_autoload}
\alias{disable_autoload}
\title{Disabling Shiny Autoload of R Scripts}
\usage{
disable_autoload(golem_wd = get_golem_wd(), pkg)
}
\arguments{
\item{pkg}{Path to the root of the package. Default is \code{get_golem_wd()}.}
}
\value{
The path to the file, invisibly.
}
\description{
Disabling Shiny Autoload of R Scripts
}
\examples{
if (interactive()) {
  disable_autoload()
}
}
