% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_deploy_helpers.R
\name{dock_from_desc}
\alias{dock_from_desc}
\title{Create Dockerfile from DESCRIPTION}
\usage{
dock_from_desc(path = "DESCRIPTION", FROM = "rocker/r-ver",
  AS = NULL, sysreqs = TRUE, repos = "https://cran.rstudio.com/",
  expand = FALSE)
}
\arguments{
\item{path}{}

\item{FROM}{}

\item{AS}{}

\item{sysreqs}{}

\item{repos}{}

\item{expand}{}
}
\description{
Create Dockerfile from DESCRIPTION
}
