% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_deploy_helpers.R
\name{add_dockerfile}
\alias{add_dockerfile}
\alias{add_dockerfile_shinyproxy}
\alias{dadd_dockerfile_heroku}
\title{Create a Dockerfile for  Shiny App}
\usage{
add_dockerfile(input = "DESCRIPTION", output = "Dockerfile",
  from = paste0("rocker/tidyverse:", R.Version()$major, ".",
  R.Version()$minor), as = NULL, port = 80, host = "0.0.0.0")

add_dockerfile_shinyproxy(input = "DESCRIPTION", output = "Dockerfile",
  from = paste0("rocker/tidyverse:", R.Version()$major, ".",
  R.Version()$minor), as = NULL)

dadd_dockerfile_heroku(input = "DESCRIPTION", output = "Dockerfile",
  from = paste0("rocker/tidyverse:", R.Version()$major, ".",
  R.Version()$minor), as = NULL)
}
\arguments{
\item{input}{path to the DESCRIPTION file to use as an input.}

\item{output}{name of the Dockerfile output.}

\item{from}{The FROM of the Dockerfile. Default is FROM rocker/tidyverse:
with `R.Version()$major` and `R.Version()$minor`.}

\item{as}{The AS of the Dockerfile. Default it NULL.}

\item{port}{The `options('shiny.port')` on which to run the Shiny App.
Default is 80.}

\item{host}{The `options('shiny.host')` on which to run the Shiny App.
Default is 0.0.0.0.}
}
\description{
Build a container containing your Shiny App. `add_dockerfile()` creates 
a "generalistic" Dockerfile, while `add_shinyproxy_dockerfile()` and 
`add_heroku_dockerfile()` creates plateform specific Dockerfile.
}
\examples{
\dontrun{
add_dockerfile()
add_shinyproxy_dockerfile()
add_heroku_dockerfile()
}
}
