% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload.R
\name{document_and_reload}
\alias{document_and_reload}
\title{Document and reload your package}
\usage{
document_and_reload(
  golem_wd = get_golem_wd(),
  roclets = NULL,
  load_code = NULL,
  clean = FALSE,
  export_all = FALSE,
  helpers = FALSE,
  attach_testthat = FALSE,
  ...,
  pkg
)
}
\arguments{
\item{golem_wd}{Path to the root of the package. Default is \code{get_golem_wd()}.}

\item{roclets}{Character vector of \link[roxygen2:roclet]{roclets} to use.

The default, \code{NULL}, uses the roxygen \code{roclets} option,
which defaults to \code{c("collate", "namespace", "rd")}. This will update
(if needed) the \code{Collate} field with \code{\link[roxygen2:update_collate]{update_collate()}},
produce the \code{NAMESPACE} file with \code{\link[roxygen2:namespace_roclet]{namespace_roclet()}}, and
produce the Rd files with \code{\link[roxygen2:rd_roclet]{rd_roclet()}}.

(Note that \code{update_collate()} is not technically a roclet but is still
controlled with this argument for historical reasons.)}

\item{load_code}{A function used to load all the R code in the package
directory. The default, \code{NULL}, uses the strategy defined by
the \code{load} roxygen option, which defaults to \code{\link[roxygen2:load_pkgload]{load_pkgload()}}.
See \link[roxygen2]{load} for more details.}

\item{clean}{If \code{TRUE}, roxygen will delete all files previously
created by roxygen before running each roclet.}

\item{export_all}{If \code{TRUE} (the default), export all objects.
If \code{FALSE}, export only the objects that are listed as exports
in the NAMESPACE file.}

\item{helpers}{if \code{TRUE} loads \pkg{testthat} test helpers.}

\item{attach_testthat}{If \code{TRUE}, attach \pkg{testthat} to the search path,
which more closely mimics the environment within test files.}

\item{...}{Other arguments passed to \code{pkgload::load_all()}}

\item{pkg}{Deprecated, please use golem_wd instead}
}
\value{
Used for side-effects
}
\description{
This function calls \code{rstudioapi::documentSaveAll()},
\code{roxygen2::roxygenise()} and \code{pkgload::load_all()}.
}
