% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload.R
\name{document_and_reload}
\alias{document_and_reload}
\title{Document and reload your package}
\usage{
document_and_reload(pkg = ".")
}
\arguments{
\item{pkg}{Path to the root of the package. Default is `"."`}
}
\description{
Document and reload your package
}
