% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enable_roxygenize.R
\name{enable_roxygenize}
\alias{enable_roxygenize}
\title{Enable documentation generation using roxygen2}
\usage{
enable_roxygenize(
  path = list.files(path = ".", pattern = "Rproj$", full.names = TRUE)[1]
)
}
\arguments{
\item{path}{to to Rproj file}
}
\description{
Enable documentation generation using roxygen2
}
