% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_favicon.R
\name{favicon}
\alias{favicon}
\title{Add favicon to your app}
\usage{
favicon(ico = "www/favicon.ico", rel = "shortcut icon")
}
\description{
This function adds the favicon from `ico` to your shiny app.
}
