% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_favicon.R
\name{use_favicon}
\alias{use_favicon}
\alias{remove_favicon}
\alias{favicon}
\title{Add a favicon to your shinyapp}
\usage{
use_favicon(path, pkg = ".")

remove_favicon(path = "inst/app/www/favicon.ico")

favicon(ico = "www/favicon.ico", rel = "shortcut icon")
}
\arguments{
\item{path}{Path to your favicon file (.ico or .png)}

\item{pkg}{Path to the root of the package. Default is \code{"."}}

\item{ico}{path to favicon file}

\item{rel}{rel}
}
\description{
This function adds the favicon from \code{ico} to your shiny app.
}
\examples{
\donttest{
if (interactive()){
  use_favicon()
  use_favicon(path='path/to/your/favicon.ico')
}
}
}
