% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_r_files.R
\name{add_fct}
\alias{add_fct}
\alias{add_utils}
\title{Add fct_ and utils_ files}
\usage{
add_fct(
  name,
  module = NULL,
  pkg = get_golem_wd(),
  open = TRUE,
  dir_create = TRUE
)

add_utils(
  name,
  module = NULL,
  pkg = get_golem_wd(),
  open = TRUE,
  dir_create = TRUE
)
}
\arguments{
\item{name}{The name of the file}

\item{module}{If not NULL, the file will be module specific in the naming (you don't need to add the leading \code{mod_})}

\item{pkg}{The working directory}

\item{open}{Should the file be opened once created?}

\item{dir_create}{Should the folder be created if it doesn't exist?}
}
\description{
These function adds files in the R/ folder
that starts either with fct_ or with utils_
}
