% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_r_files.R
\name{add_fct}
\alias{add_fct}
\alias{add_utils}
\title{Add fct_ and utils_ files}
\usage{
add_fct(
  name,
  module = NULL,
  pkg = get_golem_wd(),
  open = TRUE,
  dir_create = TRUE,
  with_test = FALSE
)

add_utils(
  name,
  module = NULL,
  pkg = get_golem_wd(),
  open = TRUE,
  dir_create = TRUE,
  with_test = FALSE
)
}
\arguments{
\item{name}{The name of the file}

\item{module}{If not NULL, the file will be module specific
in the naming (you don't need to add the leading \code{mod_}).}

\item{pkg}{Path to the root of the package. Default is \code{get_golem_wd()}.}

\item{open}{Should the created file be opened?}

\item{dir_create}{Creates the directory if it doesn't exist, default is \code{TRUE}.}

\item{with_test}{should the module be created with tests?}
}
\value{
The path to the file, invisibly.
}
\description{
These functions add files in the R/ folder
that starts either with \code{fct_} (short for function)
or with \code{utils_}.
}
