% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{fill_desc}
\alias{fill_desc}
\title{Fill your description}
\usage{
fill_desc(
  pkg_name,
  pkg_title,
  pkg_description,
  author_first_name,
  author_last_name,
  author_email,
  author_orcid = NULL,
  repo_url = NULL,
  pkg = get_golem_wd()
)
}
\arguments{
\item{pkg_name}{The name of the package}

\item{pkg_title}{The title of the package}

\item{pkg_description}{Description of the package}

\item{author_first_name}{First Name of the author}

\item{author_last_name}{Last Name of the author}

\item{author_email}{Email of the author}

\item{author_orcid}{ORCID of the author}

\item{repo_url}{URL (if needed)}

\item{pkg}{Path to look for the DESCRIPTION. Default is \code{get_golem_wd()}.}
}
\description{
Fill your description
}
