% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_current_config}
\alias{get_current_config}
\title{Get the path to the current config File}
\usage{
get_current_config(path = here::here())
}
\arguments{
\item{path}{Path to start looking for the config}
}
\description{
This function tries to guess where the golem-config file is located.
If it can't find it, this function asks the
user if they want to set the golem skeleton.
}
