% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dependencies.R
\name{get_dependencies}
\alias{get_dependencies}
\title{Return all package dependencies from current package}
\usage{
get_dependencies(path = "DESCRIPTION", dput = FALSE,
  field = c("Depends", "Imports"))
}
\arguments{
\item{path}{path to the DESCRIPTION file}

\item{dput}{if TRUE return a dput output instead of character vector}

\item{field}{DESCRIPTION fields to parse. Default is Import and Depends}
}
\description{
Return all package dependencies from current package
}
\examples{
\dontrun{
get_dependencies()
}
}
