% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dependencies.R
\name{get_dependencies}
\alias{get_dependencies}
\title{return all package dependencies from current package}
\usage{
get_dependencies(path = "DESCRIPTION", dput = TRUE, field = c("Depends",
  "Imports"))
}
\arguments{
\item{path}{path to the DESCRIPTION file}

\item{dput}{if FALSE return a vector instead of dput output}

\item{field}{DESCRIPTION fied to parse, Import and Depends by default}
}
\description{
return all package dependencies from current package
}
\examples{
\dontrun{
list_imports()
}
}
