% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem-package.R
\docType{package}
\name{golem}
\alias{golem}
\alias{golem-package}
\title{A package for building Shiny App}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Create a prod-ready shiny app with this RStudio project template.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ThinkR-open/golem}
  \item Report bugs at \url{https://github.com/ThinkR-open/golem/issues}
}

}
\author{
\strong{Maintainer}: Vincent Guyader \email{vincent@thinkr.fr} (0000-0003-0671-9270)

Authors:
\itemize{
  \item Colin Fay \email{contact@colinfay.me} (0000-0001-7343-1846)
}

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
