% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem-package.R
\docType{package}
\name{golem}
\alias{golem}
\alias{golem-package}
\title{A package for building Shiny App}
\description{
Read more about building big shiny apps at \url{https://thinkr-open.github.io/building-shiny-apps-workflow/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ThinkR-open/golem}
  \item Report bugs at \url{https://github.com/ThinkR-open/golem/issues}
}

}
\author{
\strong{Maintainer}: Vincent Guyader \email{vincent@thinkr.fr} (0000-0003-0671-9270)

Authors:
\itemize{
  \item Colin Fay \email{contact@colinfay.me} (0000-0001-7343-1846)
  \item Sébastien Rochette \email{sebastien@thinkr.fr} (0000-0002-1565-9313)
  \item Cervan Girard \email{cervan@thinkr.fr} (0000-0002-4816-4624)
}

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
