% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_golem_name}
\alias{get_golem_name}
\title{A function to return the golem name to be used elsewhere
in the package.}
\usage{
get_golem_name()
}
\value{
The name of the golem.
}
\description{
A function to return the golem name to be used elsewhere
in the package.
}
