% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dev_deps.R
\name{install_dev_deps}
\alias{install_dev_deps}
\title{Install {golem} dev dependencies}
\usage{
install_dev_deps(force_install = FALSE, ...)
}
\arguments{
\item{force_install}{If force_install is installed,
then the user is not interactively asked
to install them.}

\item{....}{further arguments passed to the install function.}
}
\description{
This function will run rlang::check_installed() on:
\itemize{
\item {usethis}
\item {pkgload}
\item {dockerfiler}
\item {devtools}
\item {roxygen2}
\item {attachment}
\item {rstudioapi}
\item {here}
\item {fs}
\item {desc}
\item {pkgbuild}
\item {processx}
\item {rsconnect}
\item {testthat}
\item {rstudioapi}
}
}
\examples{
if (interactive()) {
  install_dev_deps()
}
}
