% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_golem.R
\name{is_golem}
\alias{is_golem}
\title{Is the directory a golem-based app?}
\usage{
is_golem(pkg = getwd())
}
\arguments{
\item{pkg}{Path to the directory to check.
Defaults to the current working directory.}
}
\description{
Trying to guess if \code{pkg} is a golem-based app.
}
\examples{
is_golem()
}
