% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dev.R
\name{cat_dev}
\alias{cat_dev}
\alias{print_dev}
\alias{message_dev}
\alias{warning_dev}
\alias{browser_dev}
\title{Functions already made dev devependant}
\usage{
cat_dev(...)

print_dev(...)

message_dev(...)

warning_dev(...)

browser_dev(...)
}
\arguments{
\item{...}{\R objects (see \sQuote{Details} for the types of objects
    allowed).}
}
\description{
Functions already made dev devependant
}
