% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_tools.R
\name{pkg_name}
\alias{pkg_name}
\alias{pkg_version}
\alias{pkg_path}
\title{Package tools}
\usage{
pkg_name(golem_wd = get_golem_wd(), path)

pkg_version(golem_wd = get_golem_wd(), path)

pkg_path(golem_wd = getwd())
}
\arguments{
\item{golem_wd}{Path to use to read the DESCRIPTION}

\item{path}{Deprecated, use golem_wd instead}
}
\value{
The value of the entry in the DESCRIPTION file
}
\description{
These are functions to help you navigate
inside your project while developing
}
