% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_tools.R
\name{pkg_name}
\alias{pkg_name}
\alias{pkg_version}
\alias{pkg_path}
\title{Package tools}
\usage{
pkg_name(pkg = get_golem_wd())

pkg_version(pkg = get_golem_wd())

pkg_path()
}
\arguments{
\item{pkg}{Path to use to read the DESCRIPTION}
}
\value{
The value of the entry in the DESCRIPTION file
}
\description{
These are functions to help you navigate
inside your project while developing
}
