% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_tools.R
\name{pkg_name}
\alias{pkg_name}
\alias{pkg_version}
\alias{pkg_path}
\title{Package tools}
\usage{
pkg_name(path = ".")

pkg_version(path = ".")

pkg_path(path = ".")
}
\arguments{
\item{path}{Path to use to read the DESCRIPTION}
}
\description{
These are functions to help you navigate
inside your project while developing
}
