% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dev.R
\name{app_prod}
\alias{app_prod}
\alias{app_dev}
\title{Is the app in dev mode or prof mode?}
\usage{
app_prod()

app_dev()
}
\value{
`TRUE` or `FALSE` depending on the status of `getOption( "golem.app.prod")`
}
\description{
Is the app in dev mode or prof mode?
}
