% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rstudio_files.R
\name{add_positconnect_file}
\alias{add_positconnect_file}
\alias{add_rconnect_file}
\alias{add_rstudioconnect_file}
\alias{add_shinyappsio_file}
\alias{add_shinyserver_file}
\title{Add an app.R at the root of your package to deploy on RStudio Connect}
\usage{
add_positconnect_file(pkg = get_golem_wd(), open = TRUE)

add_rstudioconnect_file(pkg = get_golem_wd(), open = TRUE)

add_shinyappsio_file(pkg = get_golem_wd(), open = TRUE)

add_shinyserver_file(pkg = get_golem_wd(), open = TRUE)
}
\arguments{
\item{pkg}{Path to the root of the package. Default is \code{get_golem_wd()}.}

\item{open}{Should the created file be opened?}
}
\value{
The path to the file, invisibly.
}
\description{
Add an app.R at the root of your package to deploy on RStudio Connect
}
\note{
In previous versions, this function was called add_rconnect_file and add_rstudioconnect_file.

\code{add_rstudioconnect_file} is now deprecated; replace by \code{\link[=add_positconnect_file]{add_positconnect_file()}}.
}
\examples{
# Add a file for Connect
if (interactive()) {
  add_positconnect_file()
}
# Add a file for Shiny Server
if (interactive()) {
  add_shinyserver_file()
}
# Add a file for Shinyapps.io
if (interactive()) {
  add_shinyappsio_file()
}
}
