% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dev.R
\name{run_dev}
\alias{run_dev}
\title{Run run_dev.R}
\usage{
run_dev(file, pkg = pkgload::pkg_name())
}
\arguments{
\item{file}{File path to \code{run_dev.R}. Defaults to \code{R/run_dev.R}.}

\item{pkg}{Package name to run the file. Defaults to current active package.}
}
\value{
Used for side-effect
}
\description{
Run run_dev.R
}
