% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dev.R
\name{run_dev}
\alias{run_dev}
\title{Run run_dev.R}
\usage{
run_dev(file = "dev/run_dev.R", pkg = get_golem_wd())
}
\arguments{
\item{file}{File path to \code{run_dev.R}. Defaults to \code{R/run_dev.R}.}

\item{pkg}{Path to the root of the package. Default is \code{get_golem_wd()}.}
}
\value{
Used for side-effect
}
\description{
Run run_dev.R
}
