% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_check.R
\name{sanity_check}
\alias{sanity_check}
\title{Sanity check for R files in the project}
\usage{
sanity_check(golem_wd = get_golem_wd(), pkg)
}
\arguments{
\item{golem_wd}{Path to the root of the package. Default is \code{get_golem_wd()}.}

\item{pkg}{Deprecated, please use golem_wd instead}
}
\value{
A DataFrame if any of the words has been found.
}
\description{
This function is used check for any `browser()`` or commented
#TODO / #TOFIX / #BUG in the code
}
