% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_check.R
\name{sanity_check}
\alias{sanity_check}
\title{Sanity check for R files in the project}
\usage{
sanity_check()
}
\description{
This function is used check for any 'browser()' or commented
#TODO / #TOFIX / #BUG in the code
}
