% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_golem_options}
\alias{set_golem_options}
\title{{golem} options}
\usage{
set_golem_options(golem_name = pkgload::pkg_name(),
  golem_version = pkgload::pkg_version(),
  golem_wd = pkgload::pkg_path(), app_prod = FALSE)
}
\arguments{
\item{golem_name}{Name of the current golem.}

\item{golem_version}{Version of the current golem.}

\item{golem_wd}{Working directory of the current golem package.}

\item{app_prod}{Is the \code{{golem}} in prod mode?}
}
\description{
Set a series of options to be used internally by \code{{golem}}.
}
