% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{expect_shinytag}
\alias{expect_shinytag}
\alias{expect_shinytaglist}
\title{Test helpers}
\usage{
expect_shinytag(object)

expect_shinytaglist(object)
}
\arguments{
\item{object}{the object to test}
}
\value{
A testthat result
}
\description{
These functions are designed to be used inside the tests 
in your Shiny app package.
}
\examples{
expect_shinytag(1)
}
