% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{expect_shinytag}
\alias{expect_shinytag}
\alias{expect_shinytaglist}
\alias{expect_html_equal}
\title{Test helpers}
\usage{
expect_shinytag(object)

expect_shinytaglist(object)

expect_html_equal(ui, html)
}
\arguments{
\item{object}{the object to test}

\item{ui}{output of an UI function}

\item{html}{html file to compare to ui}
}
\value{
A testthat result
}
\description{
These functions are designed to be used inside the tests
in your Shiny app package.
}
\examples{
expect_shinytag(shiny::tags$span("1"))
expect_shinytaglist(shiny::tagList(1))
}
