% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{expect_shinytag}
\alias{expect_shinytag}
\alias{expect_shinytaglist}
\alias{expect_html_equal}
\alias{expect_running}
\title{Test helpers}
\usage{
expect_shinytag(object)

expect_shinytaglist(object)

expect_html_equal(ui, html, ...)

expect_running(sleep, R_path = NULL)
}
\arguments{
\item{object}{the object to test}

\item{ui}{output of an UI function}

\item{html}{deprecated}

\item{...}{arguments passed  to \code{testthat::expect_snapshot()}}

\item{sleep}{number of seconds}

\item{R_path}{path to R. If NULL, the function will try to guess where R is.}
}
\value{
A testthat result.
}
\description{
These functions are designed to be used inside the tests
in your Shiny app package.
}
\examples{
expect_shinytag(shiny::tags$span("1"))
expect_shinytaglist(shiny::tagList(1))
}
