% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_files.R
\name{use_external_js_file}
\alias{use_external_js_file}
\alias{use_external_css_file}
\title{Use Files}
\usage{
use_external_js_file(url, name, pkg = get_golem_wd(),
  dir = "inst/app/www", open = TRUE, dir_create = TRUE)

use_external_css_file(url, name, pkg = get_golem_wd(),
  dir = "inst/app/www", open = TRUE, dir_create = TRUE)
}
\arguments{
\item{url}{String representation of URL for the file to be downloaded}

\item{name}{The name of the module}

\item{pkg}{Path to the root of the package. Default is \code{"."}.}

\item{dir}{Path to the dir where the file while be created.}

\item{open}{Should the file be opened?}

\item{dir_create}{Creates the directory if it doesn't exist, default is \code{TRUE}.}
}
\description{
These functions download files from external sources and install them inside the approriate directory.
}
