% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_fn.R
\name{use_module_test}
\alias{use_module_test}
\title{Add a test file for a module}
\usage{
use_module_test(name, golem_wd = get_golem_wd(), open = TRUE, pkg)
}
\arguments{
\item{name}{The name of the module.}

\item{golem_wd}{Path to the root of the package. Default is \code{get_golem_wd()}.}

\item{open}{Should the created file be opened?}

\item{pkg}{Deprecated, please use golem_wd instead}
}
\value{
Used for side effect. Returns the path invisibly.
}
\description{
Add a test file for in module, with the new testServer structure.
}
