% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_readme.R
\name{use_readme_rmd}
\alias{use_readme_rmd}
\title{Generate a README.Rmd}
\usage{
use_readme_rmd(
  open = rlang::is_interactive(),
  pkg_name = golem::get_golem_name(),
  overwrite = FALSE,
  golem_wd = golem::get_golem_wd(),
  pkg
)
}
\arguments{
\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}

\item{pkg_name}{The name of the package}

\item{overwrite}{an optional \code{logical} flag; if \code{TRUE}, overwrite
existing \code{README.Rmd}, else throws an error if \code{README.Rmd} exists}

\item{golem_wd}{Path to the root of the package. Default is \code{get_golem_wd()}.}

\item{pkg}{Deprecated, please use golem_wd instead}
}
\value{
pure side-effect function that generates template \code{README.Rmd}
}
\description{
Generate a README.Rmd
}
