% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_recommended.R
\name{use_recommended_dep}
\alias{use_recommended_dep}
\alias{use_recommended_tests}
\title{Add recommended elements}
\usage{
use_recommended_dep(pkg = ".", recommended = c("shiny", "DT",
  "attempt", "glue", "htmltools"))

use_recommended_tests(pkg = ".")
}
\arguments{
\item{pkg}{Path to the root of the package. Default is `"."`.}

\item{recommended}{A vector of recommended packages.}
}
\description{
\describe{
  \item{use_recommended_dep}{Adds `shiny`, `DT`, `attempt`, `glue`, `golem` to dependencies}
  \item{use_recommended_tests}{Adds a test folder and copy the golem tests}
  \item{use_recommended_js}{Adds some JavaScript functions}
}
}
