% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_recommended.R
\name{use_recommended_deps}
\alias{use_recommended_deps}
\alias{use_recommended_tests}
\title{Add recommended elements}
\usage{
use_recommended_deps(pkg = ".", recommended = c("shiny", "DT",
  "attempt", "glue", "htmltools", "golem"))

use_recommended_tests(pkg = ".")
}
\arguments{
\item{pkg}{Path to the root of the package. Default is \code{"."}.}

\item{recommended}{A vector of recommended packages.}
}
\description{
\describe{
\item{use_recommended_deps}{Adds \code{shiny}, \code{DT}, \code{attempt}, \code{glue}, \code{golem}, \code{htmltools} to dependencies}
\item{use_recommended_tests}{Adds a test folder and copy the golem tests}
}
}
