% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_utils.R
\name{use_utils_ui}
\alias{use_utils_ui}
\alias{use_utils_test_ui}
\alias{use_utils_server}
\alias{use_utils_test_server}
\title{Use the utils files}
\usage{
use_utils_ui(golem_wd = get_golem_wd(), with_test = FALSE, pkg)

use_utils_test_ui(golem_wd = get_golem_wd(), pkg)

use_utils_server(golem_wd = get_golem_wd(), with_test = FALSE, pkg)

use_utils_test_ui(golem_wd = get_golem_wd(), pkg)

use_utils_test_server(golem_wd = get_golem_wd(), pkg)
}
\arguments{
\item{golem_wd}{Path to the root of the package. Default is \code{get_golem_wd()}.}

\item{with_test}{should the module be created with tests?}

\item{pkg}{Deprecated, please use golem_wd instead}
}
\value{
Used for side-effects.
}
\description{
\describe{
\item{use_utils_ui}{Copies the golem_utils_ui.R to the R folder.}
\item{use_utils_server}{Copies the golem_utils_server.R to the R folder.}
}
}
