% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_opt.R
\name{with_golem_options}
\alias{with_golem_options}
\title{Add Golem options to a Shiny App}
\usage{
with_golem_options(app, golem_opts)
}
\arguments{
\item{app}{the app object.}

\item{golem_opts}{A list of Options to be added to the app}
}
\value{
a shiny.appObj object
}
\description{
Add Golem options to a Shiny App
}
\note{
You'll probably never have to write this function
as it is included in the golem template created on
launch.
}
